* Fixed: priest talent to be spirit temporary after death. Prevent death from DoT/etc in spirit form. Stand up and stop breathing timers at gain spirit form. * Continue some cleanups related to player stand state.
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@4758 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/src/game/Player.cpp
@@ -15260,7 +15260,7 @@ void Player::ClearComboPoints()
 void Player::SetStandState(uint8 state)
 {
     WorldPacket data(SMSG_STANDSTATE_CHANGE_ACK, 1);
-    data << state;
+    data << (uint8)state;
     GetSession()->SendPacket(&data);
 }
